Crate try_blocks

An attribute-style macro that emulates unstable try blocks.

Limitations

Currently, the macros model error handling using the internal trait rather than std::ops::Try, since it has not stabilized yet. Therefore, it cannot be applied when the try blocks uses the user-defined type that implements Try manually.

The expansion of try blocks performs when the macro is invoked. Therefore, some try blocks are ignored if they are generated by other macros called later. If you want to delay the generation of try blocks, consider using try_block instead.

Examples

use std::io;
use std::net::TcpStream;

#[try_blocks]
fn fallible() {
    let result: io::Result<Vec<u8>> = try {
        let mut stream = TcpStream::connect("127.0.0.1:8000")?;

        let mut buf = vec![0u8; 1024];
        stream.read_exact(&mut buf[..])?;

        buf // tail expr is automatically wrapped
    };

    match result {
        Ok(data) => { /* ... */ },
        Err(err) => {
            eprintln!("caught an error: {}", err);
            std::process::exit(101);
        }
    }
}

use std::io;
use async_std::net::TcpStream;

#[try_blocks]
async fn fallible_async() {
    let result: io::Result<Vec<u8>> = try {
        let mut stream = TcpStream::connect("127.0.0.1:8000").await?;

        let mut buf = vec![0u8; 1024];
        stream.read_exact(&mut buf[..]).await?;

        buf
    };

    match result {
        Ok(data) => { /* ... */ },
        Err(err) => {
            eprintln!("caught an error: {}", err);
            std::process::exit(101);
        }
    }
}

use std::io;
use std::net::TcpStream;

let result: io::Result<Vec<u8>> = try_block! {
    let mut stream = TcpStream::connect("127.0.0.1:8000")?;

    let mut buf = vec![0u8; 1024];
    stream.read_exact(&mut buf[..])?;

    buf
};

match result {
    Ok(data) => { /* ... */ },
    Err(err) => {
        eprintln!("caught an error: {}", err);
        std::process::exit(101);
    }
}

Macros

try_block
Emulate try block.

Attribute Macros

try_blocks
Expand try blocks.